Credit Score develop

+++ src/main/java/com/appbackend/example/AppBackend/models/CreditScoreDTO.java
@@ -7,7 +7,7 @@ public class CreditScoreDTO {
 
 	private Integer totalCreditScore;
 	private long availableOffer;
-	private long totalExposure;
+	private Float totalExposure;
 	private long offerPerLevel;
 
 }
